requirements.txt
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
brotlipy=0.7.0=py36h8f6f2f9_1001
ca-certificates=2021.5.30=ha878542_0
certifi=2021.5.30=py36h5fab9bb_0
cffi=1.14.5=py36hc120d54_0
chardet=4.0.0=py36h5fab9bb_1
cloudpickle=1.6.0=py_0
cryptography=3.4.7=py36hb60f036_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py36_0
cytoolz=0.11.0=py36h7b6447c_0
dask-core=2.30.0=py_0
dataclasses=0.8=pyh4f3eec9_6
decorator=4.4.2=py_0
freetype=2.10.4=h5ab3b9f_0
idna=2.10=pyh9f0ad1d_0
imageio=2.9.0=py_0
intel-openmp=2020.2=254
jpeg=9b=h024ee3a_2
kiwisolver=1.2.0=py36hfd86e86_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.3.0=hdf63c60_0
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.3=h2531618_0
matplotlib-base=3.3.1=py36h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.2=he6710b0_1
networkx=2.5=py_0
ninja=1.10.2=py36hff7bd54_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py_0
openssl=1.1.1k=h7f98852_0
pillow=8.2.0=py36he98fc37_0
pip=21.0.1=py36h06a4308_0
pycparser=2.20=pyh9f0ad1d_2
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=py_0
pysocks=1.7.1=py36h5fab9bb_3
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytorch=1.7.1=py3.6_cuda10.2.89_cudnn7.6.5_0
pywavelets=1.1.1=py36h7b6447c_2
pyyaml=5.3.1=py36h7b6447c_1
pyzmq=20.0.0=py36h2b1bd32_1
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3deb0d_0
scikit-image=0.17.2=py36hdf5156a_0
scipy=1.1.0=py36hfa4b5c9_1
setuptools=52.0.0=py36h06a4308_0
six=1.15.0=pyhd3eb1b0_0
sqlite=3.35.3=hdfb4753_0
tifffile=2020.10.1=py36hdd07704_2
tk=8.6.10=hbc83047_0
toolz=0.11.1=py_0
torchaudio=0.7.2=py36
torchfile=0.1.0=py_0
torchvision=0.8.2=py36_cu102
tornado=6.0.4=py36h7b6447c_1
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.5=pyhd8ed1ab_0
visdom=0.1.8.9=0
websocket-client=0.57.0=py36h5fab9bb_4
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.3=h58526e2_3
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0